+++ src/api/middlewares/contentTypeValidation.ts
@@ -1,7 +1,7 @@
 import BaseResponse from '@/data/models/baseResponse';
 import type { NextFunction, Request, Response } from 'express';
-import { HTTP_STATUS_CODE } from '@/data/enums/httpStatusCode';
-import { RESPONSE_ERROR_MESSAGE } from '@/data/enums/responseMessage';
+import { HTTP_STATUS_CODE } from '@/data/constants/httpStatusCode';
+import { RESPONSE_ERROR_MESSAGE } from '@/data/constants/responseMessage';
 
 export const contentTypeValidationMiddleware = (request: Request, response: Response, next: NextFunction): void => {
 	if (request.headers['content-type'] !== 'application/json')
