+++ src/api/middlewares/errorHandler.ts
@@ -1,8 +1,8 @@
 import { logger } from '@/utils/logger';
 import BaseError from '@/error/baseError';
-import { ERROR_NAME } from '@/data/enums/errorName';
+import { ERROR_NAME } from '@/data/constants/errorName';
 import BaseResponse from '@/data/models/baseResponse';
-import { HTTP_STATUS_CODE } from '@/data/enums/httpStatusCode';
+import { HTTP_STATUS_CODE } from '@/data/constants/httpStatusCode';
 import type { NextFunction, Request, Response } from 'express';
 
 export const errorHandlerMiddleware = (
