+++ src/api/middlewares/notFoundHandler.ts
@@ -1,7 +1,7 @@
 import type { Request, Response } from 'express';
 import BaseResponse from '@/data/models/baseResponse';
-import { HTTP_STATUS_CODE } from '@/data/enums/httpStatusCode';
-import { RESPONSE_ERROR_MESSAGE } from '@/data/enums/responseMessage';
+import { HTTP_STATUS_CODE } from '@/data/constants/httpStatusCode';
+import { RESPONSE_ERROR_MESSAGE } from '@/data/constants/responseMessage';
 
 export const notFoundHandlerMiddleware = (request: Request, response: Response): void => {
 	response.status(HTTP_STATUS_CODE.NOT_FOUND).json(
