+++ src/api/middlewares/requestMethosValidator.ts
@@ -1,10 +1,10 @@
 import type { TRequestMethod } from '..';
 import AmbianceConfig from '@/config/appConfig';
 import BaseResponse from '@/data/models/baseResponse';
-import { API_CONFIG_KEY } from '@/data/enums/appConfig';
-import { HTTP_STATUS_CODE } from '@/data/enums/httpStatusCode';
+import { API_CONFIG_KEY } from '@/data/constants/appConfig';
+import { HTTP_STATUS_CODE } from '@/data/constants/httpStatusCode';
 import type { NextFunction, Request, Response } from 'express';
-import { RESPONSE_ERROR_MESSAGE } from '@/data/enums/responseMessage';
+import { RESPONSE_ERROR_MESSAGE } from '@/data/constants/responseMessage';
 
 export const requestMethodValidatorMiddleware = (request: Request, response: Response, next: NextFunction): void => {
 	const enabledHttpRequestMethods = AmbianceConfig.sharedInstance.api.get(API_CONFIG_KEY.ENABLED_HTTP_REQUEST_METHODS);
