+++ src/api/routes/router.ts
@@ -2,7 +2,7 @@ import type { Route } from './router.d';
 import { logger } from '@/utils/logger';
 import type { Application } from 'express';
 import AmbianceConfig from '@/config/appConfig';
-import { API_CONFIG_KEY } from '@/data/enums/appConfig';
+import { API_CONFIG_KEY } from '@/data/constants/appConfig';
 import { notFoundHandlerMiddleware } from '../middlewares/notFoundHandler';
 import { errorHandlerMiddleware } from '../middlewares/errorHandler';
