+++ src/config/appConfig.d.ts
@@ -1,4 +1,4 @@
-import type { API_CONFIG_KEY, APP_CONFIG_KEY } from '../data/enums/localObnject';
+import type { API_CONFIG_KEY, APP_CONFIG_KEY } from '../data/constants/appConfig';
 
 export type TAppConfig = {
 	readonly app: Map<APP_CONFIG_KEY, string | number>;
