+++ src/config/appConfig.ts
@@ -1,11 +1,11 @@
 import { name, version } from '../../package.json';
-import { API_CONFIG_KEY, APP_CONFIG_KEY, ENABLED_HTTP_REQUEST_METHODS } from '@/data/enums/appConfig';
+import { API_CONFIG_KEY, APP_CONFIG_KEY, ENABLED_HTTP_REQUEST_METHODS } from '@/data/constants/appConfig';
 
 let sharedInstance: AmbianceConfig | null = null;
 
 export default class AmbianceConfig {
-	public readonly api: Map<string, string | string[]> = new Map<string, string | string[]>();
 	public readonly app: Map<APP_CONFIG_KEY, string | number> = new Map<APP_CONFIG_KEY, string | number>();
+	public readonly api: Map<API_CONFIG_KEY, string | string[]> = new Map<API_CONFIG_KEY, string | string[]>();
 
 	static get sharedInstance(): AmbianceConfig {
 		if (sharedInstance === null) {
@@ -17,7 +17,7 @@ export default class AmbianceConfig {
 	constructor() {
 		this.app.set(APP_CONFIG_KEY.NAME, name);
 		this.app.set(APP_CONFIG_KEY.VERSION, version);
-		this.app.set(APP_CONFIG_KEY.PORT, process.env.PORT || 3000);
+		this.app.set(APP_CONFIG_KEY.PORT, process.env.PORT || 3010);
 		this.app.set(APP_CONFIG_KEY.ENV, process.env.NODE_ENV || 'development');
 
 		this.api.set(API_CONFIG_KEY.BASE_PATH, '/api/v1');
