+++ src/data/constants/appConfig.ts
