+++ src/error/httpError.d.ts
@@ -1,4 +1,4 @@
-import type { HTTP_STATUS_CODE } from '@/data/enums/httpStatusCode';
+import type { HTTP_STATUS_CODE } from '@/data/constants/httpStatusCode';
 
 export type THttpError = {
 	status: HTTP_STATUS_CODE;
